Add files via upload
Use the python file to generate the txt files, then to use them on the simulator, copy/paste them into wokwi.txt on the simulator (COPY/PASTE EXACTLY AS IS, IT'S GOING TO LOOK WEIRD, DO NOT BE WORRIED ABOUT IT), and run the simulation. Simulation files are also uploaded here, but you can also just find a link to it in simulation.txt (probably the easiest way), just copy/paste the text generated by the python script.